server/requirements.txt
asgiref==3.6.0
Django==4.2.1
sqlparse==0.4.4
psycopg2-binary==2.9.6
django-environ==0.10.0
gunicorn==20.1.0